+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+
+## 2.0.0 - 2020-01-13
 - Update dependencies
 - [**BC**] Require .net core `^3.1`
 - Add `AssemblyInfo`